Core tests v8
Moved SignUpViewTest (Abstract)
to _test_helpers.py so it can be available but not throw errors.
FAILED (failures=8, errors=6)
