+++ job_connect/core/tests/tests_forms.py
@@ -14,7 +14,7 @@ class CoreFormsTest(TestCase):
         form_data = {'username': 'test', 'email': 'test@test.com', 'password': 'short', 'password2': 'short'}
         form = ApplicantSignUpForm(form_data)
         self.assertFalse(form.is_valid())
-        print(form.errors.keys())  # Print the error keys
+        # print(form.errors.keys())  # Print the error keys
         self.assertIn('password', form.errors)
 
     def test_recruiter_signup_form(self):
@@ -26,7 +26,7 @@ class CoreFormsTest(TestCase):
         form_data = {'username': 'test', 'email': 'test@test.com', 'password': 'short', 'password2': 'short'}
         form = RecruiterSignUpForm(form_data)
         self.assertFalse(form.is_valid())
-        print(form.errors.keys())  # Print the error keys
+        #print(form.errors.keys())  # Print the error keys
         self.assertIn('password2', form.errors)  # Or 'password1' if that's the key
